chore(fixture): rename me_fixture.py to token_endpoint_fixture.py and update imports
